pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["JMHman <joenmh0710@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
django-environ = "^0.11.2"
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
